Fix Texture load loop on memory warnings
Solves 3 cases

1. when memory_warning happens very early, the value highest sometimes is 0.
This leads to a load loop because we are trying to ratchet down or increase
memory against 0, which would stay 0, because MEMORY_DROP_RATE and MEMORY_GAIN_RATE
are being multiplied against 0. So we set highest as 50MB (Please suggest a different value).

2. We are increasing the max_texture bytes on two conditions,
    a. when it is overLimit
    b. when highest > max_texture_bytes (this is already been doing right now)

3. overLimit calculation had a bug in texture_manager_clear_textures() function.
manager->approx_bytes_to_load will reduce when textures are cleared, but that was
not taken into account in the overLimit calculation.